HSM-286: Move platform-config.h to common
