+++ common/platform-config.h.in
